+++ pkg/manager/job_service.go
@@ -3,6 +3,7 @@ package manager
 import (
 	"context"
 	"fmt"
+	"time"
 
 	"github.com/go-jet/jet/v2/sqlite"
 	"github.com/kasuboski/mediaz/config"
@@ -19,6 +20,50 @@ type JobService struct {
 	jobStorage      storage.JobStorage
 	activityStorage storage.ActivityStorage
 	statsStorage    storage.StatisticsStorage
+}
+
+// TriggerJobRequest represents the request to manually trigger a job
+type TriggerJobRequest struct {
+	Type string `json:"type"`
+}
+
+// JobResponse represents a single job in API responses
+type JobResponse struct {
+	ID        int64     `json:"id"`
+	Type      string    `json:"type"`
+	State     string    `json:"state"`
+	CreatedAt time.Time `json:"createdAt"`
+	UpdatedAt time.Time `json:"updatedAt"`
+	Error     *string   `json:"error,omitempty"`
+}
+
+// JobListResponse represents a list of jobs in API responses
+type JobListResponse struct {
+	Jobs       []JobResponse    `json:"jobs"`
+	Count      int              `json:"count"`
+	Pagination *pagination.Meta `json:"pagination,omitempty"`
+}
+
+// toJobResponse converts a storage.Job to a JobResponse
+func toJobResponse(job *storage.Job) JobResponse {
+	return JobResponse{
+		ID:        int64(job.ID),
+		Type:      job.Type,
+		State:     string(job.State),
+		CreatedAt: *job.CreatedAt,
+		UpdatedAt: *job.UpdatedAt,
+		Error:     job.Error,
+	}
+}
+
+// isValidJobType validates that a job type string matches one of the defined JobType constants
+func isValidJobType(jobType string) bool {
+	switch JobType(jobType) {
+	case MovieIndex, MovieReconcile, SeriesIndex, SeriesReconcile, IndexerSync:
+		return true
+	default:
+		return false
+	}
 }
 
 func NewJobService(jobStorage storage.JobStorage, activityStorage storage.ActivityStorage, statsStorage storage.StatisticsStorage, cfg config.Manager, executors map[JobType]JobExecutor) *JobService {
